requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.35
pydantic==2.9.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
mercadopago==2.2.3
